(#117) Do not merge multiple SRV records into one
When attempting to find the PuppetDB node we have to fist try
PuppetDB SRV record and then the Puppet one and after that resort
to configured defaults.  First one wins.

Before this it queried all the given SRV records - both PuppetDB and
Puppet ones - and then merged the answers and sorted by SRV conventions
so you ended up with luck of the draw hitting either PuppetDB or Puppet
should both sets of SRV records exist.

Now it will prioritise the PuppetDB ones