(#101) Expose the last task status to the templates
This adds a basic status object for task results that gets associated
with every task.  It stores the success, msg, data, description and run
time of every task.  This object is quite basic now and will no doubt
extend a bit when we get to reports.

This will eventually be added to reports and such but for now to get us
going it's used to expose the last task status via templates as
previous_task.success or .msg etc

New template items:

  previous_task.success
  previous_task.description
  previous_task.msg or previous_task.message
  previous_task.data
  previous_task.runtime