(#228) Validate messages received by the Federation Broker
This validates that messages received on the Federation are things the
federation side would send.  Ie. they are not attempting to set targets
to other Federation Brokers or similar sneaky thing to attempt to hop
across Federations

It also validates the Collective side to ensure someone there is not
attempting to hop to other Collectives or Federations by forging
messages

Also fixes the paths to the files in the module