(#228) Allow connections to be traced
This removes the earlier planned record_route option and instead trace
based on the presence of the seen-by header.

Tracing is a debug aid its inconceivable that someone would reconfigure
everything to add the record_route option just to debug something, much
more likely is the case where they want to initiate a trace of a
specific request on demand.

This way the new 'mco federation trace' comand can initiate a custom
request requiring seen-by headers to be made while everything else
remains the same

As above this adds a 'mco federation trace some.node' command that will
show both Federated and Unfederated routes.

Not yet sure exactly how this output should look in the long run so the
printing part is horrific and brute force but until I know the final
desired outcome investing in writing a recursive printing alog seems a
waste of time