(#282) Adjust message versioning
This adjusts the versioning of the various protocol messages to be all
version 1 and also add protocol headers to the transport messages.

New version strings are:

   * choria:reply:1 and choria:request:1
   * choria:secure:reply:1 and choria:secure:request:1
   * choria:transport:1

Thus the whole grouping of messages represents version 1 rather than the
mish mash of v1 and v3 that all somehow combined.  Realistically it will
be impossible to version this stuff in finer level than as a group, so
this makes the whole group the same version

This has the same compatibility story as 8ebc3b158 and will ship at the
same time - Old clients can talk to new servers.  New clients cannot
talk to old servers.