(#296) Use base64 for encoding reply hashes
The hash computed for reply data was inadvertantly done using Hex
encoding, this is the only hex encoded thing in the entire protocol
while there are several other base64 encoded bits.

So for consistancy this is changed.  We do not validate this hash at
present so this has no impact