(#347) Adds run and status actions
This adds 3 actions, 2 for running and one status.

When running you can request a run and wait scenario where the agent
will wait till near the agent timeout for the command to finish, if it
has not finished yet by that time the then current status will be
returned.  When not waiting the request id is returned

Later status can be used with the requestid to gather status for any
task