(#347) Bolt Task AAA
This does an additional authorization and audit when a task is being
run, it will set the action to the task name and call the authorization
and auditing plugins.

The end result is that you can authorize using any of the usual auth
stuff individual bolt tasks.  With action policy you can do:

    policy default deny

    # site wide policies
    allow *  run_and_wait run_no_wait download task_status * *
    allow choria=rip.mcollective  choria::ls  * *

this allows anyone to call the basic actions in the agent and then have
one specific task authorized - all other tasks are denied