(#550) intermediate certificate chains
For bonus effort, this will also test SANs by moving the validation
logic to openssl, instead of trying to parse the CN out of the subject.

Contains ruby 2.4/2.5 compat for OpenSSL::SSL::Context depreciating the
use of `extra_cert_chain` - the ruby shipped with puppet 5.X is behind
a revision.

Ruby's OpenSSL doesn't make this easy, which I suspect is more to do
with OpenSSL than Ruby.  There are alternatives which reduce some of
this, like R509 but haven't investigated them yet.

* Add intermediate certificate processing - hacky :(
* Add intermediate chain certs
* Add spec tests to validate intermediate chain handling
* Add something to detect `/usr/bin/false` vs `/bin/false` because no one can agree where it goes.