(#638) Allow discovery with choria::service
MCollective is being replaced with Choria.  The mcollective::service
class was used to know if a node was configured with MCollective or not.

Now that the service is being removed from the MCollective class, we
need to check for the equivalent class in the choria module,
choria::service.

For now, look for any of those two classes, so that legacy nodes not
using Choria can continue to behave as previously.